roles y permisos

+++ resources/views/dashboard/pets/index.blade.php
@@ -9,10 +9,12 @@
 @section('content_header')
     <div class="flex justify-between items-center">
         <div class="text-lg font-bold">{{ trans('lang.list_pets') }}</div>
-        <a href="{{ route('dashboard.pets.create') }}"
-            class="bg-blue-500 hover:bg-blue-600 text-white p-2 rounded-md font-semibold px-4 ">
-            {{ trans('lang.createPet') }}
-        </a>
+        @can('dashboard.pets.create')
+            <a href="{{ route('dashboard.pets.create') }}"
+                class="bg-blue-500 hover:bg-blue-600 text-white p-2 rounded-md font-semibold px-4 ">
+                {{ trans('lang.createPet') }}
+            </a>
+        @endcan
     </div>
 @endsection
 <div class="card">
@@ -33,13 +35,13 @@
             <thead>
                 <tr>
                     <th>{{ trans('lang.pet_id') }}</th>
-                    <th>{{ trans('lang.name') }}</th> 
-                    <th>{{ trans('lang.castrated') }}</th> 
-                    <th>{{ trans('lang.lost') }}</th> 
-                    <th>{{ trans('lang.specie') }}</th> 
-                    <th>{{ trans('lang.duenio') }}</th> 
-                    <th>{{ trans('lang.updated_at') }}</th> 
-                    
+                    <th>{{ trans('lang.name') }}</th>
+                    <th>{{ trans('lang.castrated') }}</th>
+                    <th>{{ trans('lang.lost') }}</th>
+                    <th>{{ trans('lang.specie') }}</th>
+                    <th>{{ trans('lang.duenio') }}</th>
+                    <th>{{ trans('lang.updated_at') }}</th>
+
                     <th>Acciones</th>
                 </tr>
             </thead>
@@ -51,24 +53,29 @@
                         <td>{{ $pet->castrated == 1 ? trans('lang.yep') : trans('lang.nop') }}</td>
                         <td>{{ $pet->lost == 1 ? trans('lang.yep') : trans('lang.nop') }}</td>
                         <td>{{ trans('lang.' . $pet->specie) }}</td>
-                        <td>{{ $pet->user_id ? $pet->user_id : trans('lang.withoutOwner')}}</td>
+                        <td>{{ $pet->user_id ? $pet->user_id : trans('lang.withoutOwner') }}</td>
 
                         <td>{{ $pet->updated_at->diffForHumans() }}</td>
                         <td class="flex items-center justify-center">
-
-                            <a href="{{ route('dashboard.pets.show', $pet) }}">
-                                <i class="fas fa-eye text-gray-500 hover:text-gray-700 cursor-pointer"></i>
-                            </a>
-                            <a href="{{ route('dashboard.pets.edit', $pet) }}" class='btn btn-link'>
-                                <i class="fas fa-edit text-gray-500 hover:text-gray-700  cursor-pointer"></i>
-                            </a>
-                            {!! Form::open(['route' => ['dashboard.pets.destroy', $pet], 'method' => 'delete']) !!}
-                            {!! Form::button('<i class="fa fa-trash text-gray-500 hover:text-gray-700"></i>', [
-                                'type' => 'submit',
-                                'class' => '',
-                                'onclick' => "return confirm('Estás seguro que deseas eliminar a $pet->name')",
-                                ]) !!}
-                            {!! Form::close() !!}
+                            @can('dashboard.pets.show')
+                                <a href="{{ route('dashboard.pets.show', $pet) }}">
+                                    <i class="fas fa-eye text-gray-500 hover:text-gray-700 cursor-pointer"></i>
+                                </a>
+                            @endcan
+                            @can('dashboard.pets.edit')
+                                <a href="{{ route('dashboard.pets.edit', $pet) }}" class='btn btn-link'>
+                                    <i class="fas fa-edit text-gray-500 hover:text-gray-700  cursor-pointer"></i>
+                                </a>
+                            @endcan
+                            @can('dashboard.pets.destroy')
+                                {!! Form::open(['route' => ['dashboard.pets.destroy', $pet], 'method' => 'delete']) !!}
+                                {!! Form::button('<i class="fa fa-trash text-gray-500 hover:text-gray-700"></i>', [
+    'type' => 'submit',
+    'class' => '',
+    'onclick' => "return confirm('Estás seguro que deseas eliminar a $pet->name')",
+]) !!}
+                                {!! Form::close() !!}
+                            @endcan
                         </td>
                     </tr>
                 @endforeach
